requirements.txt
Flask
gunicorn
yfinance
pandas
google-genai
python-dotenv